Comments
git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2023 9c4724a3-da1e-0410-bc68-88f18da35cd4

+++ amibe/src/org/jcae/mesh/amibe/ds/MeshParameters.java
@@ -64,7 +64,7 @@ public class MeshParameters
 			epsilonProperty = "-1.0";
 			System.setProperty("org.jcae.mesh.amibe.ds.Mesh.epsilon", epsilonProperty);
 		}
-		epsilon = Double.valueOf(epsilonProperty);
+		epsilon = Double.valueOf(epsilonProperty).doubleValue();
 
 		String isotropicProperty = System.getProperty("org.jcae.mesh.Mesher.isotropic");
 		if (isotropicProperty == null)
